fix issues with `CommittableSource` benchmarks
We were not emitting the right number of observed elements before
