+++ src/Akka.Streams.Kafka.Benchmark/Benchmarks/Consumer/PlainConsumerBenchmark.cs
