+++ src/Akka.Streams.Kafka.Benchmark/Benchmarks/Micro/CommittableOffsetBatchBenchmark.cs
